Move cmdline parser utils into main.py
