+++ src/katgpucbf/fgpu/main.py
@@ -41,11 +41,19 @@ from .. import (
     DEFAULT_TTL,
     DIG_SAMPLE_BITS,
 )
-from ..main import add_common_arguments, add_recv_arguments, engine_main
+from ..main import (
+    add_common_arguments,
+    add_recv_arguments,
+    comma_split,
+    engine_main,
+    parse_dither,
+    parse_enum,
+    parse_source,
+)
 from ..mapped_array import make_vkgdr
 from ..monitor import FileMonitor, Monitor, NullMonitor
 from ..spead import DEFAULT_PORT
-from ..utils import DitherType, comma_split, parse_dither, parse_enum, parse_source
+from ..utils import DitherType
 from . import DIG_SAMPLE_BITS_VALID
 from .engine import FEngine
 from .output import NarrowbandOutput, NarrowbandOutputDiscard, NarrowbandOutputNoDiscard, WidebandOutput, WindowFunction
