+++ src/katgpucbf/xbgpu/main.py
@@ -50,11 +50,11 @@ from .. import (
     DEFAULT_PACKET_PAYLOAD_BYTES,
     DEFAULT_TTL,
 )
-from ..main import add_common_arguments, add_recv_arguments, engine_main
+from ..main import add_common_arguments, add_recv_arguments, engine_main, parse_dither, parse_source_ipv4
 from ..mapped_array import make_vkgdr
 from ..monitor import FileMonitor, Monitor, NullMonitor
 from ..spead import DEFAULT_PORT
-from ..utils import DitherType, parse_dither, parse_source_ipv4
+from ..utils import DitherType
 from .correlation import device_filter
 from .output import BOutput, XOutput
