+++ test/test_utils.py
@@ -29,11 +29,8 @@ from aiokatcp import DeviceStatus
 
 from katgpucbf.utils import (
     DeviceStatusSensor,
-    DitherType,
     TimeConverter,
     TimeoutSensorStatusObserver,
-    comma_split,
-    parse_dither,
 )
 
 
@@ -176,56 +173,3 @@ class TestTimeConverter:
         """Test :meth:`.TimeConverter.adc_to_unix`."""
         assert time_converter.adc_to_unix(0.0) == 1234567890.0
         assert time_converter.adc_to_unix(10485760.0) == 1234567890.0 + 10.0
-
-
-class TestCommaSplit:
-    """Test :func:`.comma_split`."""
-
-    def test_basic(self) -> None:
-        """Test normal usage, without optional features."""
-        assert comma_split(int)("3,5") == [3, 5]
-        assert comma_split(int)("3") == [3]
-        assert comma_split(int)("") == []
-
-    def test_bad_value(self) -> None:
-        """Test with a value that isn't valid for the element type."""
-        with pytest.raises(ValueError, match="invalid literal for int"):
-            assert comma_split(int)("3,hello")
-
-    def test_fixed_count(self) -> None:
-        """Test with a value for `count`."""
-        splitter = comma_split(int, 2)
-        assert splitter("3,5") == [3, 5]
-        with pytest.raises(ValueError, match="Expected 2 comma-separated fields, received 3"):
-            splitter("3,5,7")
-        with pytest.raises(ValueError, match="Expected 2 comma-separated fields, received 1"):
-            splitter("3")
-
-    def test_allow_single(self) -> None:
-        """Test with `allow_single`."""
-        splitter = comma_split(int, 2, allow_single=True)
-        assert splitter("3,5") == [3, 5]
-        assert splitter("3") == [3, 3]
-        with pytest.raises(ValueError, match="Expected 2 comma-separated fields, received 3"):
-            splitter("3,5,7")
-
-
-class TestParseDither:
-    """Test :func:`.parse_dither`."""
-
-    @pytest.mark.parametrize(
-        "input, output",
-        [("none", DitherType.NONE), ("uniform", DitherType.UNIFORM)],
-    )
-    def test_success(self, input: str, output: DitherType) -> None:
-        """Test with valid inputs."""
-        assert parse_dither(input) == output
-
-    @pytest.mark.parametrize("input", ["", "false", "UnIFoRM", "NONE", "default"])
-    def test_invalid(self, input: str) -> None:
-        """Test with invalid inputs."""
-        with pytest.raises(
-            ValueError,
-            match=rf"Invalid dither value {input} \(valid values are \['none', 'uniform'\]\)",
-        ):
-            parse_dither(input)
